+++ src/js/components/app/App.js
@@ -1,7 +1,6 @@
 // @flow
 import * as React from 'react';
-import Route from 'react-router-dom/Route';
-import Switch from 'react-router-dom/Switch';
+import { Route, Switch } from 'react-router-dom';
 import LayoutConnected from '../layout';
 import Home from '../home';
 import TodoManager from '../todoManager';