+++ src/js/components/layout/MenuDrawerConnected.js
@@ -1,24 +1,23 @@
 // @flow
 import { connect } from 'react-redux';
-import withRouter from 'react-router-dom/withRouter';
-import withStyles from '@material-ui/core/styles/withStyles';
+import { push } from 'connected-react-router';
 import actions from '../../app/actions';
 import states from '../../app/states';
 import MenuDrawer from './MenuDrawer';
-import styles from './styles';
 
 const mapStateToProps = state => ({
   shouldMenuLeftOpened: states.layout.shouldMenuLeftOpened(state),
   isMenuCurrentlyOpened: states.layout.isMenuCurrentlyOpened(state),
 });
 
-const mapDispatchToProps = {
-  onToggleMenu: actions.toggleMenu,
-};
+const mapDispatchToProps = dispatch => ({
+  onToggleMenu: () => dispatch(actions.toggleMenu()),
+  onRouteChange: (location: string) => dispatch(push(location)),
+});
 
 const MenuDrawerConnected = connect(
   mapStateToProps,
   mapDispatchToProps,
 )(MenuDrawer);
 
-export default withRouter(withStyles(styles.menuDrawer)(MenuDrawerConnected));
+export default MenuDrawerConnected;