+++ webpack.base.config.js
@@ -1,7 +1,5 @@
 const path = require('path');
-const webpack = require('webpack');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const autoprefixer = require('autoprefixer');
 const packageJson = require('./package.json');
 
 const srcPath = path.join(__dirname, packageJson.config.src_dir_path);
@@ -41,7 +39,11 @@ const webpackOptions = {
       },
       {
         test: /\.css$/,
-        use: [MiniCssExtractPlugin.loader, 'css-loader'],
+        use: [
+          MiniCssExtractPlugin.loader,
+          { loader: 'css-loader', options: { importLoaders: 1 } },
+          'postcss-loader',
+        ],
       },
       {
         test: /\.woff(2)?$/,
@@ -69,8 +71,8 @@ const webpackOptions = {
                 optimizationLevel: 4,
               },
               pngquant: {
-                quality: '75-90',
-                speed: 3,
+                quality: [0.65, 0.9],
+                speed: 4,
               },
             },
           },
@@ -79,36 +81,10 @@ const webpackOptions = {
     ],
   },
 
-  // Split vendors from app
-  optimization: {
-    splitChunks: {
-      cacheGroups: {
-        default: false,
-        commons: {
-          test: /[\\/]node_modules[\\/]/,
-          name: 'vendor',
-          chunks: 'all',
-        },
-      },
-    },
-  },
-
   plugins: [
-    new webpack.LoaderOptionsPlugin({
-      minimize: true,
-      debug: false,
-      options: {
-        // create vendor prefixes to maximize compatibility. Recommended by Google:
-        // https://developers.google.com/web/tools/setup/setup-buildtools#dont-trip-up-with-vendor-prefixes
-        postcss: [
-          autoprefixer({
-            browsers: ['last 2 versions'],
-          }),
-        ],
-      },
+    new MiniCssExtractPlugin({
+      filename: 'css/[name].[hash].css',
     }),
-
-    new MiniCssExtractPlugin(),
   ],
 
   // To suppress this warning when creating the vendor bundle:-